Refine signedUrl response type
The implementation only requires `signedUrl` to be present, and supports `headers` if provided.  Otherwise it passes through the S3 response you provide based on how your server or `getSignedUrl` implementation works.

This change tries to better capture this.

It was a bit strange that the other fields like `publicUrl`, `filename`, and `fileKey` were included before when they were not used in any way by the library.